C:/Program Files/Git/marks/attribute changed int to double

+++ JDBCTemplate/src/main/java/JBDL63/JDBCTemplate/model/Student.java
@@ -18,6 +18,6 @@ public class Student {
 
     private String department;
 
-    private Integer marks;
+    private Double marks;
 
 }
